make MultiTabs Client only
